Integrate root alias regressions into existing tests
